move tagSelectUtils and add tag creation
