+++ frontend/src/dialogs/concept/useCreateConceptDialog.js
@@ -4,7 +4,7 @@ import { CREATE_CONCEPT } from '../../graphql/Mutation'
 import { useDialog } from '../DialogProvider'
 import cache from '../../apollo/update'
 import TaxonomyTags from './TaxonomyTags'
-import tagSelectProps from './tagSelectUtils'
+import tagSelectProps from '../tagSelectUtils'
 
 const useCreateConceptDialog = (workspaceId, isStaff) => {
   const { openDialog } = useDialog()
